# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface AuthLogin {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  fullName: string;
  role: string;
  /** @nullable */
  regionId: number | null;
  /** @nullable */
  cityId: number | null;
  /** @nullable */
  leadershipId: number | null;
  isActive: boolean;
  canCreateLeaderUsers: boolean;
  permissions: string[];
}

export interface AuthSession {
  token: string;
  user: AuthUser;
}

export interface AuthPermission {
  id: number;
  key: string;
  label: string;
  description: string;
  category: string;
}

export interface AuthRole {
  key: string;
  label: string;
  description: string;
  userCount: number;
  permissionKeys: string[];
}

export interface AuthRbac {
  roles: AuthRole[];
  permissions: AuthPermission[];
}

export interface AuthRolePermissionsInput {
  permissionKeys: string[];
}

export interface AuthUserUpdate {
  fullName: string;
  role: string;
  /** @nullable */
  regionId?: number | null;
  /** @nullable */
  cityId?: number | null;
  /** @nullable */
  leadershipId?: number | null;
  isActive?: boolean;
  canCreateLeaderUsers?: boolean;
  /** @minLength 8 */
  password?: string;
}

export type AuthUserInput = AuthUserUpdate & {
  email: string;
  /** @minLength 8 */
  password: string;
};

export type AuthUserAdmin = AuthUser & ({
  /** @nullable */
  regionName?: string | null;
  /** @nullable */
  cityName?: string | null;
  /** @nullable */
  lastLoginAt?: string | null;
  createdAt?: string;
});

export interface Region {
  id: number;
  name: string;
  cityCount: number;
  leadershipCount: number;
}

export interface City {
  id: number;
  name: string;
  regionId: number;
  regionName: string;
  leadershipCount: number;
  deputyCount: number;
}

export interface FederalDeputy {
  id: number;
  name: string;
  isAlliance: boolean;
  leadershipCount: number;
  cityCount: number;
}

export interface Leadership {
  id: number;
  cityId: number;
  cityName: string;
  regionId: number;
  regionName: string;
  /** @nullable */
  internalRegion?: string | null;
  /** @nullable */
  articulatorId?: number | null;
  /** @nullable */
  articulatorName?: string | null;
  /** @nullable */
  coordinatorId?: number | null;
  /** @nullable */
  coordinatorName?: string | null;
  /** @nullable */
  coordinatorContact?: string | null;
  name: string;
  /** @nullable */
  leadershipContact?: string | null;
  /** @nullable */
  federalDeputyId?: number | null;
  /** @nullable */
  federalDeputyName?: string | null;
  /** @nullable */
  allianceStatus?: string | null;
  /** @nullable */
  originalFederalDeputy?: string | null;
  /** @nullable */
  religion?: string | null;
  sourceSheet: string;
  sourceRow: number;
  needsReview: boolean;
}

export interface LeadershipInput {
  cityId: number;
  /** @nullable */
  internalRegion?: string | null;
  /** @nullable */
  articulatorId?: number | null;
  /** @nullable */
  coordinatorId?: number | null;
  /** @nullable */
  coordinatorContact?: string | null;
  /** @minLength 1 */
  name: string;
  /** @nullable */
  leadershipContact?: string | null;
  /** @nullable */
  federalDeputyId?: number | null;
  /** @nullable */
  allianceStatus?: string | null;
  /** @nullable */
  originalFederalDeputy?: string | null;
  /** @nullable */
  religion?: string | null;
}

export interface LeadershipUpdate {
  cityId?: number;
  /** @nullable */
  internalRegion?: string | null;
  /** @nullable */
  articulatorId?: number | null;
  /** @nullable */
  coordinatorId?: number | null;
  /** @nullable */
  coordinatorContact?: string | null;
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  leadershipContact?: string | null;
  /** @nullable */
  federalDeputyId?: number | null;
  /** @nullable */
  allianceStatus?: string | null;
  /** @nullable */
  originalFederalDeputy?: string | null;
  /** @nullable */
  religion?: string | null;
  needsReview?: boolean;
}

export interface LeadershipList {
  items: Leadership[];
  total: number;
  page: number;
  pageSize: number;
}

export type CampaignOverviewTotals = {
  regions: number;
  cities: number;
  leaderships: number;
  reviewItems: number;
};

export type CampaignOverviewTopDeputiesItem = {
  name: string;
  leadershipCount: number;
};

export interface CampaignOverview {
  totals: CampaignOverviewTotals;
  regions: Region[];
  topDeputies: CampaignOverviewTopDeputiesItem[];
}

export type ReviewIssueSeverity = typeof ReviewIssueSeverity[keyof typeof ReviewIssueSeverity];


export const ReviewIssueSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ReviewIssueStatus = typeof ReviewIssueStatus[keyof typeof ReviewIssueStatus];


export const ReviewIssueStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface ReviewIssue {
  id: number;
  leadershipId: number;
  type: string;
  severity: ReviewIssueSeverity;
  details: string;
  status: ReviewIssueStatus;
  sourceSheet: string;
  sourceRow: number;
}

export type RegionIdParameter = number;

export type CityIdParameter = number;

export type FederalDeputyIdParameter = number;

export type GetCurrentUser200 = {
  user: AuthUser;
};

export type Logout200 = {
  success: boolean;
};

export type ListCitiesParams = {
/**
 * @minimum 1
 */
regionId?: RegionIdParameter;
/**
 * @minimum 1
 */
federalDeputyId?: FederalDeputyIdParameter;
};

export type ListLeadershipsParams = {
/**
 * @minimum 1
 */
regionId?: RegionIdParameter;
/**
 * @minimum 1
 */
cityId?: CityIdParameter;
/**
 * @minimum 1
 */
federalDeputyId?: FederalDeputyIdParameter;
search?: string;
reviewOnly?: boolean;
sortBy?: ListLeadershipsSortBy;
sortDirection?: ListLeadershipsSortDirection;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};

export type ListLeadershipsSortBy = typeof ListLeadershipsSortBy[keyof typeof ListLeadershipsSortBy];


export const ListLeadershipsSortBy = {
  name: 'name',
  city: 'city',
  region: 'region',
  deputy: 'deputy',
  status: 'status',
} as const;

export type ListLeadershipsSortDirection = typeof ListLeadershipsSortDirection[keyof typeof ListLeadershipsSortDirection];


export const ListLeadershipsSortDirection = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListReviewIssuesParams = {
status?: ListReviewIssuesStatus;
};

export type ListReviewIssuesStatus = typeof ListReviewIssuesStatus[keyof typeof ListReviewIssuesStatus];


export const ListReviewIssuesStatus = {
  open: 'open',
  resolved: 'resolved',
  all: 'all',
} as const;