# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLogin"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/rbac/roles/{role}:
    patch:
      operationId: updateRolePermissions
      tags: [access-control]
      summary: Update the permissions assigned to a role
      parameters:
        - $ref: "#/components/parameters/AuthRole"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRolePermissionsInput"
      responses:
        "200":
          description: Updated role
  /auth/users:
    get:
      operationId: listAuthUsers
      tags: [access-control]
      summary: List system users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUserAdmin"
    post:
      operationId: createAuthUser
      tags: [access-control]
      summary: Create a system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthUserInput"
      responses:
        "201":
          description: User created
  /auth/users/{id}:
    patch:
      operationId: updateAuthUser
      tags: [access-control]
      summary: Update a system user
      parameters:
        - $ref: "#/components/parameters/AuthUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthUserUpdate"
      responses:
        "200":
          description: User updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AuthRole:
      in: path
      name: role
      required: true
      schema:
        type: string
        enum: [ADMIN_GERAL, ARTICULADOR, COORDENADOR, LIDERANCA]
    AuthUserId:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AuthLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        regionId:
          type: ["integer", "null"]
        cityId:
          type: ["integer", "null"]
        leadershipId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        canCreateLeaderUsers:
          type: boolean
        permissions:
          type: array
          items:
            type: string
      required: [id, email, fullName, role, regionId, cityId, leadershipId, isActive, canCreateLeaderUsers, permissions]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
...
    AuthRolePermissionsInput:
      type: object
      properties:
        permissionKeys:
          type: array
          items:
            type: string
      required: [permissionKeys]
    AuthUserInput:
      allOf:
        - $ref: "#/components/schemas/AuthUserUpdate"
        - type: object
          properties:
            email:
              type: string
              format: email
            password:
              type: string
              minLength: 8
          required: [email, password, fullName, role]
    AuthUserUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        regionId:
          type: ["integer", "null"]
        cityId:
          type: ["integer", "null"]
        leadershipId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        canCreateLeaderUsers:
          type: boolean
        password:
          type: string
          minLength: 8
      required: [fullName, role]
    AuthUserAdmin:
      allOf:
        - $ref: "#/components/schemas/AuthUser"
        - type: object
          properties:
            regionName:
              type: ["string", "null"]
            cityName:
              type: ["string", "null"]
            lastLoginAt:
              type: ["string", "null"]
              format: date-time
            createdAt:
              type: string
              format: date-time
...