# lib/api-zod/src/generated/types/authRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthRole {
  key: string;
  label: string;
  description: string;
  userCount: number;
  permissionKeys: string[];
}