# lib/api-zod/src/generated/types/authUserAdmin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export type AuthUserAdmin = AuthUser & ({
  /** @nullable */
  regionName?: string | null;
  /** @nullable */
  cityName?: string | null;
  /** @nullable */
  lastLoginAt?: Date | null;
  createdAt?: Date;
});