# lib/api-zod/src/generated/types/authUserInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserUpdate } from './authUserUpdate';

export type AuthUserInput = AuthUserUpdate & {
  email: string;
  /** @minLength 8 */
  password: string;
};