# lib/api-zod/src/generated/types/authUserUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUserUpdate {
  fullName: string;
  role: string;
  /** @nullable */
  regionId?: number | null;
  /** @nullable */
  cityId?: number | null;
  /** @nullable */
  leadershipId?: number | null;
  isActive?: boolean;
  canCreateLeaderUsers?: boolean;
  /** @minLength 8 */
  password?: string;
}